+++ packages/riverpod/lib/src/framework/provider_base.dart
@@ -327,7 +327,10 @@ abstract class ProviderElementBase<State> implements Ref<State>, Node {
     }
 
     return state.map(
-      error: (error) => _rethrowProviderError(error.error, error.stackTrace),
+      error: (error) => throwErrorWithCombinedStackTrace(
+        error.error,
+        error.stackTrace,
+      ),
       data: (data) => data.state,
     );
   }
@@ -974,15 +977,6 @@ The provider ${_debugCurrentlyBuildingElement!.origin} modified $origin while bu
   String toString() {
     return '$runtimeType(provider: $provider, origin: $origin)';
   }
-}
-
-Never _rethrowProviderError(Object error, StackTrace stackTrace) {
-  final chain = Chain([
-    Trace.current(),
-    ...Chain.forTrace(stackTrace).traces,
-  ]).foldFrames((frame) => frame.package == 'riverpod');
-
-  Error.throwWithStackTrace(error, chain);
 }
 
 mixin OverrideWithValueMixin<State> on ProviderBase<State> {